move TransactionalMessage into Ohm namespace
